Fix Sword and Shield Raid Rewards editor mappings
Map Drop reward records to their exact unsigned fields, star values, signed table identity, den usage, item labels, and kind-specific limits while preserving existing full-width source values.

Harden reward FlatBuffer and GFPAK parsing, source-preserving copy-on-write output, alias isolation, malformed-layout rejection, atomic batch staging, semantic source binding, reviewed-plan fingerprints, combined Drop and Bonus rollback, and verified output replacement.

Keep desktop search, filtered selection, drafts, and prior session state stable after rejection. Add focused desktop, format, workflow, randomizer, and bridge regressions, plus test-only integration access for injected combined-apply failures.

+++ src/KM.SwSh/Raids/SwShRaidRewardsWorkflowModels.cs
@@ -31,17 +31,17 @@ public sealed record SwShRaidRewardEditableField(
 }
 
 public sealed record SwShRaidRewardEditableFieldOption(
-    int Value,
+    long Value,
     string Label);
 
 public sealed record SwShRaidRewardItemRecord(
     int Slot,
-    int EntryId,
-    int ItemId,
+    long EntryId,
+    long ItemId,
     string ItemName,
-    int Quantity,
-    int Weight,
-    IReadOnlyList<int> Values);
+    long Quantity,
+    long Weight,
+    IReadOnlyList<long> Values);
 
 public sealed record SwShRaidRewardTableRecord(
     string TableId,
